chore(parser): move strip namespace to openapi
