+++ src/openApi/common/parser/__tests__/stripNamespace.spec.ts
