+++ src/openApi/common/parser/stripNamespace.ts
