+++ src/openApi/common/parser/type.ts
@@ -1,6 +1,6 @@
-import { stripNamespace } from '../../../utils/stripNamespace';
 import type { Type } from '../interfaces/Type';
 import { sanitizeTypeName } from './sanitize';
+import { stripNamespace } from './stripNamespace';
 
 /**
  * Get mapped type for given type to basic Typescript/Javascript type.
